+++ packages/@baldeweg/ui/src/components/BSwitch/BSwitch.vue
@@ -1,23 +1,23 @@
 <script setup>
 defineProps({
-  label: {
-    type: String,
-  },
   modelValue: {
     type: Boolean,
     default: false,
   },
+  label: {
+    type: String,
+    required: true,
+  },
 })
+
+const emits = defineEmits(['update:modelValue'])
 </script>
 
 <template>
   <div class="switch">
-    <button
-      class="switch_indicator"
-      :class="{ isActive: modelValue }"
-      @click="$emit('update:modelValue', !modelValue)"
-    />
-    <div class="switch_label" @click="$emit('update:modelValue', !modelValue)">
+    <button class="switch_indicator" :class="{ isActive: modelValue }" @click="emits('update:modelValue', !modelValue)"
+      :aria-label="label" />
+    <div class="switch_label" @click="emits('update:modelValue', !modelValue)">
       {{ label }}
     </div>
   </div>
@@ -28,11 +28,13 @@ defineProps({
   display: flex;
   align-items: center;
 }
+
 .switch_label {
   display: block;
   margin-left: 10px;
   cursor: pointer;
 }
+
 .switch_indicator {
   position: relative;
   border: 0 hidden;
@@ -43,6 +45,7 @@ defineProps({
   padding: 0;
   cursor: pointer;
 }
+
 .switch_indicator::after {
   content: '';
   position: absolute;
@@ -54,9 +57,11 @@ defineProps({
   height: 20px;
   transition: 0.2s;
 }
+
 .switch_indicator.isActive {
   background: var(--color-primary-10);
 }
+
 .switch_indicator.isActive::after {
   left: calc(100% - 5px);
   transform: translateX(-100%);